Fixed virtual time increments
We can afford to increment system time V using LINK_RATE only when there
are no eligible and ready groups. If there are no eligible groups, then
V has to progress at the LINK_RATE in order to make some rate limited
groups eligible.

When there are some eligible and ready groups, we do not have to
increase V for any additional delays in processing. For example, if
another process gets scheduled on the dequeue core or if the CPU is not
able to keep up with line rate, then incrementing V just based on real
time would be bad. This would affect the bounds on grp->S and V causing
V to progress far ahead of grp->S even when the group is already
eligible and hasn't been served. This in turn could have caused problems
while trying to bucket sort a newly class that is activated in that
group which has a small value of S relative to V.